move v1 tests into v1 folder
