+++ tests/api/original/auth/test_require_auth.py
